requirements.txt
danlp==0.0.12
psycopg2_binary==2.8.6
# tqdm==4.42.1
# spacy==2.2.3
# gensim==3.8.1
# torch==1.6.0
# flair==0.4.5
# pyconll==2.2.1
# pandas==1.0.1
# transformers==3.5.1
# srsly==1.0.2
allennlp==1.3.0
# tweepy==3.10.0
# conllu==4.4